requirement.txt
Flask==3.0.3
flask-cors==5.0.0
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
